Moved baggage Unit testcases to Integration testcases
