+++ src/Propagation/Instana/phpunit.xml.dist
@@ -39,6 +39,9 @@
         <testsuite name="unit">
             <directory>tests/Unit</directory>
         </testsuite>
+        <testsuite name="integration">
+            <directory>tests/Integration</directory>
+        </testsuite>
     </testsuites>
 
 </phpunit>
